+++ src/Facades/LaravelDrafts.php
@@ -7,6 +7,13 @@ use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Facades\Facade;
 
 /**
+ * @method static \Illuminate\Contracts\Auth\Authenticatable getCurrentUser()
+ * @method static void previewMode(bool $previewMode = true)
+ * @method static void disablePreviewMode()
+ * @method static bool isPreviewModeEnabled()
+ * @method static void withDrafts(bool $withDrafts = true)
+ * @method static bool isWithDraftsEnabled()
+ *
  * @see \Oddvalue\LaravelDrafts\LaravelDrafts
  * @method Model | Authenticatable getCurrentUser();
  */
@@ -14,6 +21,6 @@ class LaravelDrafts extends Facade
 {
     protected static function getFacadeAccessor(): string
     {
-        return 'laravel-drafts';
+        return \Oddvalue\LaravelDrafts\LaravelDrafts::class;
     }
 }