+++ CHANGELOG.md
@@ -2,6 +2,31 @@
 
 All notable changes to `laravel-drafts` will be documented in this file.
 
+## v1.2.0 - 2023-02-22
+
+### What's Changed
+
+- New feature: preview mode by @oddvalue in https://github.com/oddvalue/laravel-drafts/pull/19
+- Add support for Laravel 10 by @oddvalue in https://github.com/oddvalue/laravel-drafts/pull/20
+
+**Full Changelog**: https://github.com/oddvalue/laravel-drafts/compare/v1.1.0...v1.2.0
+
+## v1.1.0 - 2023-02-20
+
+### What's Changed
+
+- Added a method to avoid creating revison by @Froxz in https://github.com/oddvalue/laravel-drafts/pull/15
+
+### New Contributors
+
+- @Froxz made their first contribution in https://github.com/oddvalue/laravel-drafts/pull/15
+
+**Full Changelog**: https://github.com/oddvalue/laravel-drafts/compare/v1.0.2...v1.1.0
+
+## v1.0.2 - 2023-02-17
+
+**Full Changelog**: https://github.com/oddvalue/laravel-drafts/compare/v1.0.1...v1.0.2
+
 ## v0.0.3 - 2022-07-01
 
 **Full Changelog**: https://github.com/oddvalue/laravel-drafts/compare/v0.0.2...v0.0.3